fix(writer): avoid flushing while forking (#15599)
## Description

Fix a race condition where `_exporter.send` between the execution of
`stop_workers` in the before fork hook and the actual fork takes place.
Since `send` recreates the tokio runtime if it has been stopped, this
undoes the `stop_workers` call and causes the writer to deadlock in
child.

Also fixes a potential deadlock happening when attempting to lock the
`_request` Event in periodic service `stop` after a fork.

## Testing

This has been tested locally in an airflow cluster which has been shown
to deadlock.

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->

---------

Co-authored-by: Brett Langdon <brett.langdon@datadoghq.com>